sign: tweak interface
