+++ hyperdrive/packages/sign/pkg/manifest.json
