--- hyperdrive/packages/signer/api/sign:sys-v0.wit
@@ -1,23 +1,0 @@
-interface sign{
-  use standard.{address};
-
-  // takes blob for message to sign or verify
-    variant request{
-        sign,
-        verify(verify-request)
-    }
-  // takes blob
-    record verify-request{
-        node: string,
-        signature: list<u8>
-    }
-  // takes blob
-    record sign-response{
-        signature: list<u8> 
-    }
-}
-
-world sign-sys-v0 {
-    import sign;
-    include process-v1;
-}
